+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Features
 
+- CLI write a more verbose message when unhandled exceptions are raised.
 - `--dump-events` output reduced by hiding element metadata and compacting
   location data.
 - use jest snapshots to test element metadata as it is more maintainable.